Refactor: #69 UserSettingPage에서 이미지, 링크 관련 코드를 컴포넌트로 분리
